+++ assets/javascripts/kitten/components/form/input/tag-input/index.js
@@ -7,7 +7,6 @@ import COLORS from '../../../../constants/colors-config'
 import TYPOGRAPHY from '../../../../constants/typography-config'
 import { CrossIcon } from '../../../graphics/icons/cross-icon'
 import { ScreenConfig } from '../../../../constants/screen-config'
-import { checkDeprecatedSizes } from '../../../../helpers/utils/deprecated'
 
 export const StyledTagInputWrapper = styled.div`
   position: relative;
@@ -137,7 +136,6 @@ export const StyledTagInputWrapper = styled.div`
 
   // Sizes
 
-  &.k-Form-TagInput--tiny,
   &.k-Form-TagInput--small {
     --tagInput-padding-vertical: ${pxToRem(5)};
     --tagInput-padding-horizontal: ${pxToRem(8)};
@@ -145,12 +143,10 @@ export const StyledTagInputWrapper = styled.div`
     min-height: ${pxToRem(40)};
   }
 
-  &.k-Form-TagInput--regular,
   &.k-Form-TagInput--medium {
     min-height: ${pxToRem(50)};
   }
 
-  &.k-Form-TagInput--big,
   &.k-Form-TagInput--large {
     min-height: ${pxToRem(60)};
     --tagInput-tag-height: ${pxToRem(32)};
@@ -195,8 +191,6 @@ export const TagInput = ({
   disabled,
   size,
 }) => {
-  checkDeprecatedSizes(size)
-
   const [itemsList, setItemsList] = useState(
     Array.from(itemsListFromProps || initialItemsList),
   )