+++ assets/javascripts/kitten/components/form/select/dropdown-select/styles.js
@@ -189,7 +189,6 @@ export const StyledDropdown = styled.div`
   /* SIZES */
 
   &.k-Form-Dropdown--micro,
-  &.k-Form-Dropdown--tiny,
   &.k-Form-Dropdown--small {
     .k-Form-DropdownCombobox__statusBadges,
     .k-Form-DropdownSelect__button__statusBadges {
@@ -204,17 +203,14 @@ export const StyledDropdown = styled.div`
     --DropdownSelect-buttonHeight: ${pxToRem(30)};
   }
 
-  &.k-Form-Dropdown--tiny,
   &.k-Form-Dropdown--small {
     --DropdownSelect-buttonHeight: ${pxToRem(40)};
   }
 
-  &.k-Form-Dropdown--normal,
   &.k-Form-Dropdown--medium {
     --DropdownSelect-buttonHeight: ${pxToRem(50)};
   }
 
-  &.k-Form-Dropdown--big,
   &.k-Form-Dropdown--large,
   &.k-Form-Dropdown--huge,
   &.k-Form-Dropdown--giant {
@@ -433,9 +429,7 @@ export const StyledDropdown = styled.div`
       }
     }
 
-    &.k-Form-Dropdown--normal,
     &.k-Form-Dropdown--medium,
-    &.k-Form-Dropdown--big,
     &.k-Form-Dropdown--large,
     &.k-Form-Dropdown--huge,
     &.k-Form-Dropdown--giant {