+++ assets/javascripts/kitten/components/structure/cards/reward-card/index.js
@@ -3,7 +3,6 @@ import PropTypes from 'prop-types'
 import classNames from 'classnames'
 import { StyledRewardCard } from './styles'
 import {
-  Info,
   Form,
   Title,
   Image,
@@ -89,7 +88,6 @@ export const RewardCard = ({
   )
 }
 
-RewardCard.Info = Info
 RewardCard.Form = Form
 RewardCard.Title = Title
 RewardCard.Image = Image